backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.3.0
langchain==0.2.14
langchain-core==0.2.38
langchain-groq==0.1.9
langgraph==0.2.4
httpx==0.27.0
